Implement processForm internally
Instead of asking the microservice writer to process the form values
directly, with access to the request and response writer, retrieve the
values from the form and attach them to the job object.

Microservice writer will instead only need to implement the function
that the job runs given the form values.  It will also need access to
GIN clients to make calls to the GIN service.