Bot and user authentication
Service logs in as a bot user with a username and password provided
through configuration (currently a file in the cwd).
The client for this bot user is attached to the worker queue.

Another client is created on behalf of the user for each job in the
queue.  This is attached to a new UserJob object that contains a
reference to the old (db mapped) db.Job type and a reference to the user
client.  The client is created in the processForm() function during the
preparation of the job.

The signature of the user-defined JobAction function now also contains
two gogs.Client arguments, one for the bot and one for the user.  The
custom function should use these to perform administrative or user
actions.