Attach username to new custom client
GOGS Client struct doesn't contain the username it belongs to.  Also,
the GOGS API endpoint for retrieving a user's own info isn't exposed in
the client.  It is impossible to retrieve the user's own info unless a
custom well call is made to the "/user" endpoint with the stored token.
This is a bit of a hassle, so instead we get around this by creating our
own client in the worker package that include the username associated
with the client.  The username is also stored in the database alongside
the session info.

This change should be reverted when the GOGS Client finally adds the
function to retrieve a user's own info.