Update GOGS client and user GetSelfInfo()
I initially stored the username in the Client struct because there was
no way to get the user's info without knowing the username beforehand.
A new method in the GOGS client now allows calling the "self" info API
endpoint, so we don't need to store it anymore.

I'm keeping the worker.Client around for future abstractions and
convenience methods.