update huwebdesign (#49)
* go mod tidy

* utonics/labproject: set local user before pushing

If a global user isn't configured, the warning will show up as an error
in the log. Set the local repository user to the tonic bot user name
before pushing.

* utonics/labproject: support reading branch info for submodules

Submodules can also specify a tracking branch. Read it while parsing
.gitmodules if it's specified.

* utonics/labproject: update .gitmodules file with new URLs

Create relative URLs for the new submodules and write them back to the
.gitmodules file. This makes working with submodules much more flexible
since it works for both git and http remotes, depending on how the
parent repository is cloned.

It also gets around any issues with the URLs in the GIN web interface.

* utonics/labproject: add .main suffix to parent repo

Add the suffix to .main to the parent repository and sanitise all
repository names so they don't include '/'.

Create admin user with 'admin' rights instead of just 'write'.

* utonics/labproject: use gin client internals to init

* utonics/labproject: link each submodule README to parent repo

* utonics/labproject: support adding common submodule

Not used yet.

* utonics/labproject: add labcommon submodule conditionally

Only add the labcommon submodule if it exists.

* utonics/labproject: add paths argument to commit

* utonics/labproject: change common submodule name to 'labcommons'

Add messages to log.

* utonics/labproject: update submodules after modifying READMEs

* utonics/labproject: clone the commons submodule

Clone the commons submodule after adding its configuration to the
.gitmodules file and add its path to the next commit.

* utonics/labproject: push changes in main repository

After modifying the submodules and updating the .gitmodules file, push
again.

* utonics/labproject: don't clone commons submodule if not found

If the commons submodule isn't found, don't clone it.
Also, don't commit until after all the submodules have been updated
(after the README).

* utonics/labproject: run git submodule add to get commons

* utonics/labproject: use proper commons URL for cloning

It will be replaced later when we write the .gitmodules file.

* change path for labcommons submodule

* erase deprecated message

---------

Co-authored-by: Achilleas Koutsou <achilleas@koutsou.net>